backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pydantic>=1.0,<2
psycopg2-binary>=2.9.0